move function into evaluator.py
